Upgrade Sentry to latest
Sentry recommends calling Sentry.init() as early as possible, so it has
been moved to the top of the entry file.
