+++ package.json
@@ -111,8 +111,7 @@
     "@microsoft/applicationinsights-web": "^3.3.8",
     "@ministryofjustice/frontend": "^5.1.5",
     "@playwright/test": "^1.52.0",
-    "@sentry/node": "^7.14.1",
-    "@sentry/tracing": "^7.120.3",
+    "@sentry/node": "^10.11.0",
     "@total-typescript/shoehorn": "^0.1.0",
     "accessible-autocomplete": "^3.0.1",
     "agentkeepalive": "^4.6.0",
